@blegat 2017-09-18
Do you think that the implementation of the evaluation in DynamicPolynomials could be improved ?