@saschatimme 2017-09-18
I don't think so (at least I don't see an obvious way to improve things).

This new algorithm relies heavily on [these precomputations](https://github.com/saschatimme/FixedPolynomials.jl/blob/a5ec57dd5f41217aab073fe701bb61702bec5ddb/src/poly.jl#L36-L60) which I would have to be recomputed for each new polynomial. So I think we cannot have the cake and eat it too :/